requirement.txt
allure-pytest==2.15.0
allure-python-commons==2.15.0
attrs==25.3.0
certifi==2025.8.3
charset-normalizer==3.4.2
DateTime==5.5
distlib==0.4.0
et_xmlfile==2.0.0
exceptiongroup==1.3.0
filelock==3.18.0
greenlet==3.2.3
idna==3.10
iniconfig==2.1.0
numpy==2.0.2
openpyxl==3.1.5
packaging==25.0
pandas==2.3.1
platformdirs==4.3.8
playwright==1.54.0
pluggy==1.6.0
pyee==13.0.0
Pygments==2.19.2
pytest==8.4.1
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.4
six==1.17.0
tomli==2.2.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
virtualenv==20.33.0
VirtualEnvironment==0.1.2
zope.interface==7.2
